Adding scripts from Helene & Colin for plotting outputs from a single site run.
Helene and Colin wrote these, and I simply cleaned them up, and have added
support for command line arguments. Trying to generalize the scripts so
that they work "out of the box"

I am not completely happy with the way way the locations of generated files
end up..we'll have to play with this stuff and see if it would be better to
have all the visulaization code in a sub-directort of the model itself.