Change from a "scrolling window" type plot to one that grows from the left.
This is a totally different approach in that now I don't use a deque
and instead actaully check the year and month tags in the json object.

If the yr and month are w/in range, then it sets the data int the plot line.
Otherwise, it ignores the data. So the plot ends up with "holes" in it for
pass-thru files that it misses (i.e. if the plotting script is slower than
the model generates data)