Initial commit of script for plotting cmtXXX_XXX-XX.nc files.
This script may replace the R scripts (output-extract.R and
plotting-output.R). Instead of generating the intermediate
csv files, this script plots the netcdf files directly. Right
now there is some support for stitching together different run-
stages, as well as support for changing which PFT to plot. The
script works on yly and mly files. There is minimal error checking
and overall I expect that it should get heavily refactored. But it
is a start.